Time: 0 ms (100.00%) | Memory: 7.9 MB (16.37%) - LeetSync

+++ 70-climbing-stairs/climbing-stairs.cpp
@@ -5,8 +5,8 @@ public:
             return 1;
         else if (n == 2)
             return 2;
-        long unsigned int last = 2, beforeLast = 1, current;
-        for (long unsigned int i = 3; i <= n; i++) {
+         int last = 2, beforeLast = 1, current;
+        for ( int i = 3; i <= n; i++) {
             current = last + beforeLast;
             beforeLast = last;
             last = current;
